requirements.txt
absl-py==2.3.1
alembic==1.17.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.2
coverage==7.11.0
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.9.23
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.11
ml_dtypes==0.5.3
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
protobuf==3.20.3
psycopg2-binary==2.9.11
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rich==14.2.0
robotframework==7.1.1
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.6.0
scipy==1.15.3
selenium==4.25.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
SQLAlchemy==2.0.44
tomli==2.3.0
trio==0.31.0
trio-websocket==0.12.2
typer==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
